# GODRIVER-3119 bson/primitive: improve DateTime and ObjectID JSON performance