GODRIVER-3119

## Summary

Improve the JSON marshal/unmarshal performance of the DateTime and ObjectID types by around 4-18x and significantly reduce allocations. This change also brings parity to the behavior of  `ObjectID.UnmarshalText` and `ObjectID.UnmarshalJSON`, which is important because the encoding/json the will use `UnmarshalText` instead of `UnmarshalJSON` when decoding a string.

```
goos: darwin
goarch: arm64
pkg: go.mongodb.org/mongo-driver/bson/primitive
                             │  base.10.txt  │             new.10.txt              │
                             │    sec/op     │   sec/op     vs base                │
ObjectIDJSON/Marshal-10        114.900n ± 1%   6.012n ± 0%  -94.77% (p=0.000 n=10)
ObjectIDJSON/Unmarshal-10       265.70n ± 3%   21.32n ± 0%  -91.98% (p=0.000 n=10)
ObjectIDJSON/UnmarshalOID-10     439.3n ± 2%   401.5n ± 1%   -8.60% (p=0.000 n=10)
DateTimeMarshalJSON-10          257.20n ± 0%   60.99n ± 0%  -76.29% (p=0.000 n=10)
DateTimeUnmarshalJSON-10        218.10n ± 1%   41.65n ± 0%  -80.90% (p=0.000 n=10)
geomean                          237.3n        42.00n       -82.30%

                             │ base.10.txt  │                new.10.txt                │
                             │     B/s      │      B/s        vs base                  │
ObjectIDJSON/Marshal-10        215.8Mi ± 1%    4123.9Mi ± 0%  +1810.90% (p=0.000 n=10)
ObjectIDJSON/Unmarshal-10      93.31Mi ± 3%   1162.95Mi ± 0%  +1146.30% (p=0.000 n=10)
ObjectIDJSON/UnmarshalOID-10   75.97Mi ± 2%     83.13Mi ± 1%     +9.42% (p=0.000 n=10)
DateTimeMarshalJSON-10         96.41Mi ± 0%    406.53Mi ± 0%   +321.68% (p=0.000 n=10)
DateTimeUnmarshalJSON-10       113.7Mi ± 1%     595.3Mi ± 0%   +423.68% (p=0.000 n=10)
geomean                        110.9Mi          626.5Mi        +464.94%

                             │ base.10.txt │               new.10.txt                │
                             │    B/op     │    B/op     vs base                     │
ObjectIDJSON/Marshal-10         72.00 ± 0%    0.00 ± 0%  -100.00% (p=0.000 n=10)
ObjectIDJSON/Unmarshal-10       200.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ObjectIDJSON/UnmarshalOID-10    552.0 ± 0%   256.0 ± 0%   -53.62% (p=0.000 n=10)
DateTimeMarshalJSON-10         104.00 ± 0%   48.00 ± 0%   -53.85% (p=0.000 n=10)
DateTimeUnmarshalJSON-10        168.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
geomean                         169.3                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                             │ base.10.txt │               new.10.txt                │
                             │  allocs/op  │ allocs/op   vs base                     │
ObjectIDJSON/Marshal-10         3.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ObjectIDJSON/Unmarshal-10       4.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ObjectIDJSON/UnmarshalOID-10    8.000 ± 0%   6.000 ± 0%   -25.00% (p=0.000 n=10)
DateTimeMarshalJSON-10          3.000 ± 0%   1.000 ± 0%   -66.67% (p=0.000 n=10)
DateTimeUnmarshalJSON-10        2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                         3.565                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```

## Background & Motivation

For fun and faster is always better.