@charlievieth 2024-01-26
**NB:** Getting the following test error which I need to address before this PR should be reviewed (missing it before pushing):
```
--- FAIL: TestObjectIdJSONMarshaling (0.00s)
    --- FAIL: TestObjectIdJSONMarshaling/1 (0.00s)
        bson_test.go:1703:
            	Error Trace:	/Users/charlie.vieth/go/src/go.mongodb.org/mongo-driver/bson/mgocompat/bson_test.go:1703
            	Error:      	Expected value not to be nil.
            	Test:       	TestObjectIdJSONMarshaling/1
            	Messages:   	expected a unmarshal error
FAIL
FAIL	go.mongodb.org/mongo-driver/bson/mgocompat	0.565s
```